Update Minimum Remove to Make Valid Parentheses
When you see in problem formulation, that it is something about valid parantheses, you should immedietly thing about stack, because the very classical problems in this topic uses stack.

Let us traverse our string from left to right and make sure, that we do not have any problems with balance:

    If symbol is (, then we increase balance by 1 and add this symbol to ans.
    If symbol is ), then we can add this symbol only if balance is positive: if it is 0, then we can not recover our string if we continue.
    If symbol not ( and not ), we add it to ans.

However, it is not enough to do only one traverse: so far when we finished, we can say that:

    Balance is never negative
    But balance in the end should be equal to 0.

These 2 conditions and sufficient and enough to have valid parentheses and at the moment only the fist one is fulfilled. What we can do now is to traverse string from the end and keep removing symbols until balance becomes equal to 0. Or we can use our clean function and apply it to reversed string.

Complexity: time complexity is O(n), we traverse our string twice. Space complexity is also O(n).