Broken Calculator
On a broken calculator that has a number showing on its display, we can perform two operations:

    Double: Multiply the number on the display by 2, or;
    Decrement: Subtract 1 from the number on the display.

Initially, the calculator is displaying the number X.

Return the minimum number of operations needed to display the number Y.

 

Example 1:

Input: X = 2, Y = 3
Output: 2
Explanation: Use double operation and then decrement operation {2 -> 4 -> 3}.

Example 2:

Input: X = 5, Y = 8
Output: 2
Explanation: Use decrement and then double {5 -> 4 -> 8}.

Example 3:

Input: X = 3, Y = 10
Output: 3
Explanation:  Use double, decrement and double {3 -> 6 -> 5 -> 10}.

Example 4:

Input: X = 1024, Y = 1
Output: 1023
Explanation: Use decrement operations 1023 times.

 

Note:

    1 <= X <= 10^9
    1 <= Y <= 10^9

Let us consider several cases and understand, what will be the best choise in each of them. Let us go in opposite direction and divide Y by 2 or add 1, to it until we get X.

    If X > Y, we do not have a lot of choice, we can just decrease X by one until it becomes equal to Y, so answer will be X - Y.
    If X == Y, then we already happy and we return 0.
    If Y % 2 == 1, then let us think, what can be the last step? It can not be multilication by 2, so the only choice is subtraction of 1 and on previous step we have configuration (X, Y + 1), for which we run our function recursively.
    If Y % 2 == 0, let us prove that we always need to divide by 2 in this case.

Imagine, that we have sequence of operations, like: [+1, +1, /2, /2, /2, +1, /2, ...]. Then, if we have +1, +1, /2 sequence inside, we can always replace it with /2, +1, so we can make it shorter. So, there will be no +1, +1 subsequence, except for the very end and in general all sequence looks like: [+1, /2, .., /2, +1, /2, ..., /2, ..., +1, ..., +1]. Note, that the last part correspondes to case 1. Also we will never have two +1 in the middle and it means, that if we have even number, we must divide it by 2: if we add 1 to it, then we have no choice to add 1 again (case 3) and then we have +1, +1, /2 pattern.

Complexity: time complexity is O(log Y), because every two iteration number decreased at least twice. Space complexity here is also O(log Y), because we use recursion. It can be made O(1) if we do it iteratively.