Advantage Shuffle
Given two arrays A and B of equal size, the advantage of A with respect to B is the number of indices i for which A[i] > B[i].

Return any permutation of A that maximizes its advantage with respect to B.

 

Example 1:

Input: A = [2,7,11,15], B = [1,10,4,11]
Output: [2,11,7,15]

Example 2:

Input: A = [12,24,8,32], B = [13,25,32,11]
Output: [24,32,8,12]

 

Note:

    1 <= A.length = B.length <= 10000
    0 <= A[i] <= 10^9
    0 <= B[i] <= 10^9

The idea here is to sort A and B lists first and start with the biggest number from B and check if we can beat it with biggest number from A. If we can, do it and increase beg pointer. If no, than we need to choose the smallest number we have and we move end pointer. One difficulty here is that we need to keep order of data somehow, so we add indexes to B list. Then we create ans list and change corresponding element of it.

Time complexity is O(n log n) to perform sorts and space complexity is O(n).