Flip Binary Tree To Match Preorder Traversal
You are given the root of a binary tree with n nodes, where each node is uniquely assigned a value from 1 to n. You are also given a sequence of n values voyage, which is the desired pre-order traversal of the binary tree.

Any node in the binary tree can be flipped by swapping its left and right subtrees. For example, flipping node 1 will have the following effect:

Flip the smallest number of nodes so that the pre-order traversal of the tree matches voyage.

Return a list of the values of all flipped nodes. You may return the answer in any order. If it is impossible to flip the nodes in the tree to make the pre-order traversal match voyage, return the list [-1].

 

Example 1:

Input: root = [1,2], voyage = [2,1]
Output: [-1]
Explanation: It is impossible to flip the nodes such that the pre-order traversal matches voyage.

Example 2:

Input: root = [1,2,3], voyage = [1,3,2]
Output: [1]
Explanation: Flipping node 1 swaps nodes 2 and 3, so the pre-order traversal matches voyage.

Example 3:

Input: root = [1,2,3], voyage = [1,2,3]
Output: []
Explanation: The tree's pre-order traversal already matches voyage, so no nodes need to be flipped.

 

Constraints:

    The number of nodes in the tree is n.
    n == voyage.length
    1 <= n <= 100
    1 <= Node.val, voyage[i] <= n
    All the values in the tree are unique.
    All the values in voyage are unique.

For me it was not medium problem, more like hard. First of all it is asked in problem statement to find the minimum number of flips, however in fact we do not have choice: I think everything is determenistic here. The first thing you need to think about when you see tree, it is dfs or bfs. Let us choose recursive dfs, it is usally easier to code and understand.

    self.ans is our list with flipped nodes.
    self.ind is where exaclty at our voyage array we are now. This is tricky part! If you do not use something similar, you still can solve problem, but it will be more like O(n^2) complexity.

dfs(node) function will work recursively as usual.

    Check if we go outside our tree or we out of elements in voyage, in this case we just go back.
    Also, we expect, that value of current node we are in now is equal to element we are currently in voyage: if it is not the case, no flipping can save this situation. So, if we see that this condition does not hold, we add None to the answer and return. In the end we going to check if we have None in answer or not.
    Define dr = 1: direction of traversal (means left -> right and -1 means right -> left), also increase self.ind by one.
    Check if we need to do flip: we need if node.left.val is not equal to voyage[self.ind]: in this case what is expected to be next node in traversal is not equal to the next node in voyages. We add node to answer and change direction. Note, that in the opposite case we can not flip this node, that is everything is determenistic.
    Traverse children in right direction and run dfs(child).
    In the end, run dfs(root) and check if we have None inside, if we have it means that at one moment of time we have impossible option and we need to return [-1]. In opposite case we return self.ans.

Complexity: What we did here is just traversed our tree, using dfs, so time complexity is O(n), where n is number of nodes in tree. Space complexity is potentially O(n) as well.